Split out tests for humanreadable into their own module, so we can port just humanreadable.py in isolation.
